requirements.txt
Flask==1.1.2
psycopg2==2.8.6
SQLAlchemy==1.4.14
pytest==6.2.4
Werkzeug==0.16.0
flask-restplus==0.13.0
python-dotenv==0.17.1
mysql-connector-python==8.0.25
PyJWT==2.1.0

